Delete.js added which has the actual mutations
